+++ src/assets/svgs/bottomNav/IcDeliver.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgIcDeliver = (props: SVGProps<SVGSVGElement>) => (
   <svg
