+++ src/assets/svgs/bottomNav/IcLocation.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 const SvgIcLocation = (props: SVGProps<SVGSVGElement>) => (
   <svg
